root/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
sqlparse==0.5.0
tzdata==2024.1
